Script to combine and extract Author Information
This script uses the information on authors extracted from GROBID in XML files as well as scraped from PDFs. Furthermore, it extracts OSF ids for the authors, ORCIDs based on OSF ids, from the PDF and from names. Then, it uses ORCIDs to check for email addresses and adds email addresses scraped from the PDF.

The script produces a csv file `authorList_ext.csv` containing all the information for the authors. Here are the columns of this table: 

* `id` : preprint OSF id
* `name.surname` : author surname from the XML
* `name.given` : author given names from the XML
* `email` : author email
* `affiliation` : author affiliation from the XML
* `n` : author number based on XML
* `orcid.xml` : ORCID based on XML, if present, otherwise NA
* `source` : source of the preprint, e.g. OSF or psyarxiv
* `osf.id` : OSF id of the author
* `osf.name` : full name based on the OSF id
* `orcid.osf` : ORCID based on OSF id
* `orcid.pdf` : ORCID scraped from the PDF
* `name.given.orcid` : given name based on the ORCID scraped from the PDF
* `name.surname.orcid` : surname based on the ORCID scraped from the PDF
* `orcid.name` : ORCID based on the name - only if exact match
* `email.source` : source of the email in the `email` column
* `orcid` : ORCID, based on all the sources for ORCIDs
* `orcid.source` :  source of the ORCID in the `orcid` column
* `pdf.email` : emails scraped from the PDF, not assigned to a specific author

A sample of 499 preprints with 2407 authors led to 88% of the preprints have at least one email associated with it, but only 26% of the authors have an ORCID and 47% have an OSF id.